+++ kubernetes/terraform/gke/backend.tf
@@ -1,0 +1,6 @@
+terraform {
+  backend "gcs" {
+    bucket = "kubernetes-tf-state-bucket-20181022001"
+    prefix = "gke/terraform/state"
+  }
+}