+++ kubernetes/terraform/gke/outputs.tf
@@ -1,0 +1,4 @@
+
+output "kubernetes_endpoint" {
+  value = "${module.gke.kubernetes_endpoint}"
+}