+++ kubernetes/terraform/modules/gke/outputs.tf
@@ -1,0 +1,4 @@
+
+output "kubernetes_endpoint" {
+  value = "${google_container_cluster.kubernetes.endpoint}"
+}